Test issues solved
